+++ src/OutputFormat.php
@@ -176,7 +176,7 @@ class OutputFormat
     {
         $aVarPrefixes = ['a', 's', 'm', 'b', 'f', 'o', 'c', 'i'];
         foreach ($aVarPrefixes as $sPrefix) {
-            $sFieldName = $sPrefix . ucfirst($sName);
+            $sFieldName = $sPrefix . \ucfirst($sName);
             if (isset($this->$sFieldName)) {
                 return $this->$sFieldName;
             }
@@ -193,20 +193,20 @@ class OutputFormat
     public function set($aNames, $mValue)
     {
         $aVarPrefixes = ['a', 's', 'm', 'b', 'f', 'o', 'c', 'i'];
-        if (is_string($aNames) && strpos($aNames, '*') !== false) {
+        if (\is_string($aNames) && \strpos($aNames, '*') !== false) {
             $aNames =
                 [
-                    str_replace('*', 'Before', $aNames),
-                    str_replace('*', 'Between', $aNames),
-                    str_replace('*', 'After', $aNames),
+                    \str_replace('*', 'Before', $aNames),
+                    \str_replace('*', 'Between', $aNames),
+                    \str_replace('*', 'After', $aNames),
                 ];
-        } elseif (!is_array($aNames)) {
+        } elseif (!\is_array($aNames)) {
             $aNames = [$aNames];
         }
         foreach ($aVarPrefixes as $sPrefix) {
             $bDidReplace = false;
             foreach ($aNames as $sName) {
-                $sFieldName = $sPrefix . ucfirst($sName);
+                $sFieldName = $sPrefix . \ucfirst($sName);
                 if (isset($this->$sFieldName)) {
                     $this->$sFieldName = $mValue;
                     $bDidReplace = true;
@@ -230,12 +230,12 @@ class OutputFormat
      */
     public function __call($sMethodName, array $aArguments)
     {
-        if (strpos($sMethodName, 'set') === 0) {
-            return $this->set(substr($sMethodName, 3), $aArguments[0]);
-        } elseif (strpos($sMethodName, 'get') === 0) {
-            return $this->get(substr($sMethodName, 3));
-        } elseif (method_exists(OutputFormatter::class, $sMethodName)) {
-            return call_user_func_array([$this->getFormatter(), $sMethodName], $aArguments);
+        if (\strpos($sMethodName, 'set') === 0) {
+            return $this->set(\substr($sMethodName, 3), $aArguments[0]);
+        } elseif (\strpos($sMethodName, 'get') === 0) {
+            return $this->get(\substr($sMethodName, 3));
+        } elseif (\method_exists(OutputFormatter::class, $sMethodName)) {
+            return \call_user_func_array([$this->getFormatter(), $sMethodName], $aArguments);
         } else {
             throw new \Exception('Unknown OutputFormat method called: ' . $sMethodName);
         }
@@ -248,7 +248,7 @@ class OutputFormat
      */
     public function indentWithTabs($iNumber = 1)
     {
-        return $this->setIndentation(str_repeat("\t", $iNumber));
+        return $this->setIndentation(\str_repeat("\t", $iNumber));
     }
 
     /**
@@ -258,7 +258,7 @@ class OutputFormat
      */
     public function indentWithSpaces($iNumber = 2)
     {
-        return $this->setIndentation(str_repeat(" ", $iNumber));
+        return $this->setIndentation(\str_repeat(' ', $iNumber));
     }
 
     /**
@@ -274,10 +274,7 @@ class OutputFormat
         return $this->oNextLevelFormat;
     }
 
-    /**
-     * @return void
-     */
-    public function beLenient()
+    public function beLenient(): void
     {
         $this->bIgnoreExceptions = true;
     }
@@ -303,10 +300,8 @@ class OutputFormat
 
     /**
      * Creates an instance of this class without any particular formatting settings.
-     *
-     * @return self
      */
-    public static function create()
+    public static function create(): OutputFormat
     {
         return new OutputFormat();
     }
@@ -319,8 +314,8 @@ class OutputFormat
     public static function createCompact()
     {
         $format = self::create();
-        $format->set('Space*Rules', "")
-            ->set('Space*Blocks', "")
+        $format->set('Space*Rules', '')
+            ->set('Space*Blocks', '')
             ->setSpaceAfterRuleName('')
             ->setSpaceBeforeOpeningBrace('')
             ->setSpaceAfterSelectorSeparator('')