+++ src/Parsing/UnexpectedTokenException.php
@@ -43,7 +43,7 @@ class UnexpectedTokenException extends SourceException
         } elseif ($this->sMatchType === 'identifier') {
             $sMessage = "Identifier expected. Got “{$sFound}”";
         } elseif ($this->sMatchType === 'custom') {
-            $sMessage = trim("$sExpected $sFound");
+            $sMessage = \trim("$sExpected $sFound");
         }
 
         parent::__construct($sMessage, $iLineNo);