+++ src/Property/Selector.php
@@ -14,11 +14,11 @@ class Selector
      * @var string
      */
     private const NON_ID_ATTRIBUTES_AND_PSEUDO_CLASSES_RX = '/
-        (\.[\w]+)                   # classes
+        (\\.[\\w]+)                   # classes
         |
-        \[(\w+)                     # attributes
+        \\[(\\w+)                     # attributes
         |
-        (\:(                        # pseudo classes
+        (\\:(                         # pseudo classes
             link|visited|active
             |hover|focus
             |lang
@@ -38,9 +38,9 @@ class Selector
      * @var string
      */
     private const ELEMENTS_AND_PSEUDO_ELEMENTS_RX = '/
-        ((^|[\s\+\>\~]+)[\w]+   # elements
+        ((^|[\\s\\+\\>\\~]+)[\\w]+   # elements
         |
-        \:{1,2}(                # pseudo-elements
+        \\:{1,2}(                    # pseudo-elements
             after|before|first-letter|first-line|selection
         ))
         /ix';
@@ -55,9 +55,9 @@ class Selector
     public const SELECTOR_VALIDATION_RX = '/
         ^(
             (?:
-                [a-zA-Z0-9\x{00A0}-\x{FFFF}_^$|*="\'~\[\]()\-\s\.:#+>]* # any sequence of valid unescaped characters
-                (?:\\\\.)?                                              # a single escaped character
-                (?:([\'"]).*?(?<!\\\\)\2)?                              # a quoted text like [id="example"]
+                [a-zA-Z0-9\\x{00A0}-\\x{FFFF}_^$|*="\'~\\[\\]()\\-\\s\\.:#+>]* # any sequence of valid unescaped characters
+                (?:\\\\.)?                                                     # a single escaped character
+                (?:([\'"]).*?(?<!\\\\)\\2)?                                    # a quoted text like [id="example"]
             )*
         )$
         /ux';
@@ -79,7 +79,7 @@ class Selector
      */
     public static function isValid($sSelector)
     {
-        return preg_match(static::SELECTOR_VALIDATION_RX, $sSelector);
+        return \preg_match(static::SELECTOR_VALIDATION_RX, $sSelector);
     }
 
     /**
@@ -107,16 +107,13 @@ class Selector
      *
      * @return void
      */
-    public function setSelector($sSelector)
+    public function setSelector($sSelector): void
     {
-        $this->sSelector = trim($sSelector);
+        $this->sSelector = \trim($sSelector);
         $this->iSpecificity = null;
     }
 
-    /**
-     * @return string
-     */
-    public function __toString()
+    public function __toString(): string
     {
         return $this->getSelector();
     }
@@ -130,9 +127,9 @@ class Selector
             $a = 0;
             /// @todo should exclude \# as well as "#"
             $aMatches = null;
-            $b = substr_count($this->sSelector, '#');
-            $c = preg_match_all(self::NON_ID_ATTRIBUTES_AND_PSEUDO_CLASSES_RX, $this->sSelector, $aMatches);
-            $d = preg_match_all(self::ELEMENTS_AND_PSEUDO_ELEMENTS_RX, $this->sSelector, $aMatches);
+            $b = \substr_count($this->sSelector, '#');
+            $c = \preg_match_all(self::NON_ID_ATTRIBUTES_AND_PSEUDO_CLASSES_RX, $this->sSelector, $aMatches);
+            $d = \preg_match_all(self::ELEMENTS_AND_PSEUDO_ELEMENTS_RX, $this->sSelector, $aMatches);
             $this->iSpecificity = ($a * 1000) + ($b * 100) + ($c * 10) + $d;
         }
         return $this->iSpecificity;