+++ src/Value/CalcFunction.php
@@ -19,22 +19,17 @@ class CalcFunction extends CSSFunction
     private const T_OPERATOR = 2;
 
     /**
-     * @param ParserState $oParserState
-     * @param bool $bIgnoreCase
-     *
-     * @return CalcFunction
-     *
      * @throws UnexpectedTokenException
      * @throws UnexpectedEOFException
      */
-    public static function parse(ParserState $oParserState, $bIgnoreCase = false)
+    public static function parse(ParserState $oParserState, bool $bIgnoreCase = false): CSSFunction
     {
         $aOperators = ['+', '-', '*', '/'];
         $sFunction = $oParserState->parseIdentifier();
         if ($oParserState->peek() != '(') {
             // Found ; or end of line before an opening bracket
             throw new UnexpectedTokenException('(', $oParserState->peek(), 'literal', $oParserState->currentLine());
-        } elseif (!in_array($sFunction, ['calc', '-moz-calc', '-webkit-calc'])) {
+        } elseif (!\in_array($sFunction, ['calc', '-moz-calc', '-webkit-calc'], true)) {
             // Found invalid calc definition. Example calc (...
             throw new UnexpectedTokenException('calc', $sFunction, 'literal', $oParserState->currentLine());
         }
@@ -65,7 +60,7 @@ class CalcFunction extends CSSFunction
                 $oCalcList->addListComponent($oVal);
                 $iLastComponentType = CalcFunction::T_OPERAND;
             } else {
-                if (in_array($oParserState->peek(), $aOperators)) {
+                if (\in_array($oParserState->peek(), $aOperators, true)) {
                     if (($oParserState->comes('-') || $oParserState->comes('+'))) {
                         if (
                             $oParserState->peek(1, -1) != ' '
@@ -84,10 +79,10 @@ class CalcFunction extends CSSFunction
                     $iLastComponentType = CalcFunction::T_OPERATOR;
                 } else {
                     throw new UnexpectedTokenException(
-                        sprintf(
+                        \sprintf(
                             'Next token was expected to be an operand of type %s. Instead "%s" was found.',
-                            implode(', ', $aOperators),
-                            $oVal
+                            \implode(', ', $aOperators),
+                            $oParserState->peek()
                         ),
                         '',
                         'custom',