+++ src/Value/Size.php
@@ -71,23 +71,21 @@ class Size extends PrimitiveValue
     /**
      * @param bool $bIsColorComponent
      *
-     * @return Size
-     *
      * @throws UnexpectedEOFException
      * @throws UnexpectedTokenException
      */
-    public static function parse(ParserState $oParserState, $bIsColorComponent = false)
+    public static function parse(ParserState $oParserState, $bIsColorComponent = false): Size
     {
         $sSize = '';
         if ($oParserState->comes('-')) {
             $sSize .= $oParserState->consume('-');
         }
-        while (is_numeric($oParserState->peek()) || $oParserState->comes('.') || $oParserState->comes('e', true)) {
+        while (\is_numeric($oParserState->peek()) || $oParserState->comes('.') || $oParserState->comes('e', true)) {
             if ($oParserState->comes('.')) {
                 $sSize .= $oParserState->consume('.');
             } elseif ($oParserState->comes('e', true)) {
                 $sLookahead = $oParserState->peek(1, 1);
-                if (is_numeric($sLookahead) || $sLookahead === '+' || $sLookahead === '-') {
+                if (\is_numeric($sLookahead) || $sLookahead === '+' || $sLookahead === '-') {
                     $sSize .= $oParserState->consume(2);
                 } else {
                     break; // Reached the unit part of the number like "em" or "ex"
@@ -100,8 +98,8 @@ class Size extends PrimitiveValue
         $sUnit = null;
         $aSizeUnits = self::getSizeUnits();
         foreach ($aSizeUnits as $iLength => &$aValues) {
-            $sKey = strtolower($oParserState->peek($iLength));
-            if (array_key_exists($sKey, $aValues)) {
+            $sKey = \strtolower($oParserState->peek($iLength));
+            if (\array_key_exists($sKey, $aValues)) {
                 if (($sUnit = $aValues[$sKey]) !== null) {
                     $oParserState->consume($iLength);
                     break;
@@ -116,17 +114,17 @@ class Size extends PrimitiveValue
      */
     private static function getSizeUnits()
     {
-        if (!is_array(self::$SIZE_UNITS)) {
+        if (!\is_array(self::$SIZE_UNITS)) {
             self::$SIZE_UNITS = [];
-            foreach (array_merge(self::ABSOLUTE_SIZE_UNITS, self::RELATIVE_SIZE_UNITS, self::NON_SIZE_UNITS) as $val) {
-                $iSize = strlen($val);
+            foreach (\array_merge(self::ABSOLUTE_SIZE_UNITS, self::RELATIVE_SIZE_UNITS, self::NON_SIZE_UNITS) as $val) {
+                $iSize = \strlen($val);
                 if (!isset(self::$SIZE_UNITS[$iSize])) {
                     self::$SIZE_UNITS[$iSize] = [];
                 }
-                self::$SIZE_UNITS[$iSize][strtolower($val)] = $val;
+                self::$SIZE_UNITS[$iSize][\strtolower($val)] = $val;
             }
 
-            krsort(self::$SIZE_UNITS, SORT_NUMERIC);
+            \krsort(self::$SIZE_UNITS, SORT_NUMERIC);
         }
 
         return self::$SIZE_UNITS;
@@ -134,10 +132,8 @@ class Size extends PrimitiveValue
 
     /**
      * @param string $sUnit
-     *
-     * @return void
      */
-    public function setUnit($sUnit)
+    public function setUnit($sUnit): void
     {
         $this->sUnit = $sUnit;
     }
@@ -153,7 +149,7 @@ class Size extends PrimitiveValue
     /**
      * @param float|int|string $fSize
      */
-    public function setSize($fSize)
+    public function setSize($fSize): void
     {
         $this->fSize = (float) $fSize;
     }
@@ -179,20 +175,17 @@ class Size extends PrimitiveValue
      *
      * @return false if the unit an angle, a duration, a frequency or the number is a component in a Color object.
      */
-    public function isSize()
+    public function isSize(): bool
     {
-        if (in_array($this->sUnit, self::NON_SIZE_UNITS, true)) {
+        if (\in_array($this->sUnit, self::NON_SIZE_UNITS, true)) {
             return false;
         }
         return !$this->isColorComponent();
     }
 
-    /**
-     * @return bool
-     */
-    public function isRelative()
+    public function isRelative(): bool
     {
-        if (in_array($this->sUnit, self::RELATIVE_SIZE_UNITS, true)) {
+        if (\in_array($this->sUnit, self::RELATIVE_SIZE_UNITS, true)) {
             return true;
         }
         if ($this->sUnit === null && $this->fSize != 0) {
@@ -201,24 +194,18 @@ class Size extends PrimitiveValue
         return false;
     }
 
-    /**
-     * @return string
-     */
-    public function __toString()
+    public function __toString(): string
     {
         return $this->render(new OutputFormat());
     }
 
-    /**
-     * @return string
-     */
-    public function render(OutputFormat $oOutputFormat)
+    public function render(OutputFormat $oOutputFormat): string
     {
-        $l = localeconv();
-        $sPoint = preg_quote($l['decimal_point'], '/');
-        $sSize = preg_match("/[\d\.]+e[+-]?\d+/i", (string) $this->fSize)
-            ? preg_replace("/$sPoint?0+$/", "", sprintf("%f", $this->fSize)) : $this->fSize;
-        return preg_replace(["/$sPoint/", "/^(-?)0\./"], ['.', '$1.'], $sSize)
-            . ($this->sUnit === null ? '' : $this->sUnit);
+        $l = \localeconv();
+        $sPoint = \preg_quote($l['decimal_point'], '/');
+        $sSize = \preg_match('/[\\d\\.]+e[+-]?\\d+/i', (string) $this->fSize)
+            ? \preg_replace("/$sPoint?0+$/", '', \sprintf('%f', $this->fSize)) : $this->fSize;
+        return \preg_replace(["/$sPoint/", '/^(-?)0\\./'], ['.', '$1.'], $sSize)
+            . ($this->sUnit ?? '');
     }
 }