+++ CHANGELOG.md
@@ -6,26 +6,52 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
-
+- Support arithmetic operators in CSS function arguments (#607)
+- Add support for inserting an item in a CSS list (#545)
 - Add a class diagram to the README (#482)
 - Add support for the `dvh`, `lvh` and `svh` length units (#415)
 - Add more tests (#449)
 
 ### Changed
 
-- Improve performance of Value::parseValue with many delimiters by refactoring to remove array_search()
+- Improve performance of Value::parseValue with many delimiters by refactoring to remove array_search() (#413)
 - Add visibility to all class/interface constants (#469)
 
 ### Deprecated
 
+- Deprecate `DeclarationBlock::createBorderShorthand()` (#578)
+- Deprecate `DeclarationBlock::createFontShorthand()` (#580)
+- Deprecate `DeclarationBlock::createDimensionsShorthand()` (#579)
+- Deprecate `DeclarationBlock::createListStyleShorthand()` (#577)
+- Deprecate `DeclarationBlock::createBackgroundShorthand()` (#576)
+- Deprecate `DeclarationBlock::createShorthandProperties()` (#575)
+- Deprecate `DeclarationBlock::expandListStyleShorthand()` (#574)
+- Deprecate `DeclarationBlock::expandBackgroundShorthand()` (#573)
+- Deprecate `DeclarationBlock::expandFontShorthand()` (#572)
+- Deprecate `DeclarationBlock::expandDimensionsShorthand()` (#571)
+- Deprecate `DeclarationBlock::expandBorderShorthand()` (#570)
+- Deprecate `DeclarationBlock::createShorthands()` (#569)
+- Deprecate `Document::expandShorthands()` (#566)
+- Deprecate `Document::createShorthands()` (#567)
+- Deprecate `DeclarationBlock::expandShorthands()` (#558)
+
 ### Removed
 
+- Remove `Rule::getValues()` (#582)
+- Remove `Rule::setValues()` (#562)
+- Remove `Document::getAllSelectors()` (#561)
+- Remove `DeclarationBlock::getSelector()` (#559)
+- Remove `DeclarationBlock::setSelector()` (#560)
 - Drop support for PHP < 7.2 (#420)
 
 ### Fixed
 
+- Fix undefined local variable in `CalcFunction::parse()` (#593)
 - Fix PHP notice caused by parsing invalid color values having less than 6 characters (#485)
 - Fix (regression) failure to parse at-rules with strict parsing (#456)
+
+@ziegenberg is a new contributor to this release and did a lot of the heavy
+lifting. Thanks! :heart:
 
 ## 8.5.0